requirements.txt
fabric
patchwork
boto3
PyYAML
sqlite3
cron